+++ __tests__/input.js
@@ -4,6 +4,7 @@ import { imageSnapshotOptions, timeouts } from './constants.json';
 
 import allOutgoingActivitiesSent from './setup/conditions/allOutgoingActivitiesSent';
 import minNumActivitiesShown from './setup/conditions/minNumActivitiesShown.js';
+import uiConnected from './setup/conditions/uiConnected';
 
 // selenium-webdriver API doc:
 // https://seleniumhq.github.io/selenium/docs/api/javascript/module/selenium-webdriver/index_exports_WebDriver.html
@@ -108,6 +109,45 @@ test('textarea long text', async () => {
   await textarea.sendKeys(
     'https://github.com/microsoft/BotFramework-WebChat/blob/master/packages/component/src/Styles/defaultStyleOptions.js'
   );
+
+  const base64PNG = await driver.takeScreenshot();
+
+  expect(base64PNG).toMatchImageSnapshot(imageSnapshotOptions);
+});
+
+test('text box input background color', async () => {
+  const props = {
+    styleOptions: {
+      sendBoxBackground: '#0063B1',
+      sendBoxTextColor: 'White',
+      sendBoxButtonColor: 'White',
+      sendBoxPlaceholderColor: 'White'
+    }
+  };
+
+  const { driver } = await setupWebDriver({ props });
+
+  await driver.wait(uiConnected(), timeouts.directLine);
+
+  const base64PNG = await driver.takeScreenshot();
+
+  expect(base64PNG).toMatchImageSnapshot(imageSnapshotOptions);
+});
+
+test('text area input background color', async () => {
+  const props = {
+    styleOptions: {
+      sendBoxBackground: '#0063B1',
+      sendBoxTextColor: 'White',
+      sendBoxButtonColor: 'White',
+      sendBoxPlaceholderColor: 'White',
+      sendBoxTextWrap: true
+    }
+  };
+
+  const { driver } = await setupWebDriver({ props });
+
+  await driver.wait(uiConnected(), timeouts.directLine);
 
   const base64PNG = await driver.takeScreenshot();